Basic working implementation of music.
The notes/and generations are generated as discussed, to the best of my
understanding.

To get the score looking even reasonable, I disabled \relative mode, as
we really want absolute mode. Relative mode is designed for manual
input, absolute is better for programatically generated music.

Sounds reasonable on first play though, seems to work as expected \o/